Make clear separation between Ca and ClusterCa classes
